Update xcon.py
Optimized the process shutdown command to have multiple flags in a singular command and will raise a warning if the command is wrong and added an !t flag to process shutdown, which specifies the time it'll take until process shutdown initiates. There has also been a few tweaks and highlighting fixes done to the [...] help command.