Update xcon.py
A new update has been made to XCon. This includes @variable tracking to explicitly mention that you're using a variable that will get parsed as the variable's value after in the command, this is purely for commands that don't focus on variable recognition (for example, you can make a variable named x, for example and use it in commands like "inspect folder @x". There has also been an escape sequence added for comments to create an actual dollar sign ($) instead of a comment, which is $$. Aside from that, there are now two new commands: command history (!r) and command search, these are documented in both the `help` command and the `utilities help`.